Adding intvalue

+++ src/main/java/com/skunkworks/spout/YoutubeSpout.java
@@ -72,9 +72,9 @@ public class YoutubeSpout extends BaseRichSpout{
 				 vids.setId(ids);
 				 List<Video> videos = vids.execute().getItems();
 				 for(Video video:videos)
-					 _collector.emit(new Values(video.getId() + "," + video.getSnippet().getTitle(), video.getStatistics().getViewCount()
-							 , video.getStatistics().getLikeCount(), video.getStatistics().getDislikeCount(),  
-							 video.getStatistics().getCommentCount(), video.getStatistics().getFavoriteCount()));
+					 _collector.emit(new Values(video.getId() + "," + video.getSnippet().getTitle(), video.getStatistics().getViewCount().intValue()
+							 , video.getStatistics().getLikeCount().intValue(), video.getStatistics().getDislikeCount().intValue(),  
+							 video.getStatistics().getCommentCount().intValue(), video.getStatistics().getFavoriteCount().intValue()));
 			 }
 		} catch (IOException e) {
 			e.printStackTrace();
